Added SandboxSettings
